refactor(artifacts): share study text validation
